Implemented generators
Now uses generator functions for creating different "objects", and collections of pointers to managers for easier parameter definitions.

+++ EntityAntFarm/input_system.cpp
@@ -2,15 +2,15 @@
 
 void input_system(olc::PixelGameEngine* engine, ScreenOffset& offset, const float fElapsedTime) {
 	if (engine->GetKey(olc::Key::W).bHeld) {
-		offset.yOffset += 100*fElapsedTime;
+		offset.yOffset += static_cast<int>(100*fElapsedTime);
 	}
 	if (engine->GetKey(olc::Key::S).bHeld) {
-		offset.yOffset -= 100 * fElapsedTime;
+		offset.yOffset -= static_cast<int>(100 * fElapsedTime);
 	}
 	if (engine->GetKey(olc::Key::A).bHeld) {
-		offset.xOffset += 100 * fElapsedTime;
+		offset.xOffset += static_cast<int>(100 * fElapsedTime);
 	}
 	if (engine->GetKey(olc::Key::D).bHeld) {
-		offset.xOffset -= 100 * fElapsedTime;
+		offset.xOffset -= static_cast<int>(100 * fElapsedTime);
 	}
 }
